Added support for multiple workers
TemporalModule now extends ConfigurableModuleBuilder, which isolates the options provider to the single instance of the module and allows for multiple invocations of the module (and multiple worker creations). The registerWorker and registerWorkerAsync methods have been added to clarify the exact purpose of the now-deprecated forRoot and forRootAsync methods, and hint to users that these methods need not be called solely for imports to AppModule.